requirements.txt
streamlit
pymongo
openai
google-generativeai
anthropic
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
pandas 